#212 - cleanup more code
