+++ angular-client/compose.client.yml
@@ -6,7 +6,7 @@ services:
     build:
       context: .
       args:
-        PROD: 'true'
+        PROD: "true"
         BACKEND_URL: http://192.168.100.1:8000
         MAP_ACCESS_TOKEN: pk.eyJ1IjoibWNrZWVwIiwiYSI6ImNscXBrcmU1ZTBscWIya284cDFyYjR3Nm8ifQ.6TQHlxhAJzptZyV-W28dnw
       target: production
