--- angular-client/src/components/connection-display/connection-display.component.css
@@ -1,6 +1,0 @@
-.connection-dot {
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  margin-top: 20px;
-}
