--- angular-client/src/components/connection-display/connection-display.component.html
@@ -1,8 +1,0 @@
-<info-background svgIcon="cell_tower" title="Connection">
-  <div style="display: flex; justify-content: center; align-items: center; height: 80%">
-    <vstack>
-      <div class="connection-dot" [style.background-color]="getConnectedColor(connected)"></div>
-      <typography variant="info-subtitle" [content]="getConnectedStatus(connected)" />
-    </vstack>
-  </div>
-</info-background>
