--- angular-client/src/components/viewer-display/viewer-display.component.html
@@ -1,7 +1,0 @@
-<info-background svgIcon="eye" title="Viewers">
-  <typography
-    style="display: flex; align-content: center; justify-content: center; height: 80%"
-    variant="info-value"
-    [content]="numViewers.toString()"
-  />
-</info-background>
