+++ angular-client/src/pages/landing-page/components/connection-display/connection-display.component.css
