+++ angular-client/src/pages/landing-page/components/date-location-display/date-location.component.ts
@@ -1,28 +1,39 @@
-import { Component, HostListener, OnInit } from '@angular/core';
+import { Component, HostListener } from '@angular/core';
 import { MessageService } from 'primeng/api';
 import { startNewRun } from 'src/api/run.api';
 import APIService from 'src/services/api.service';
 import Storage from 'src/services/storage.service';
 import { IdentifierDataType } from 'src/utils/enumerations/identifier-data-type';
 
-/**
- * Container for the landing page, obtains data from the storage service.
- */
 @Component({
-  selector: 'landing-page-header',
-  styleUrls: ['./landing-page-header.component.css'],
-  templateUrl: './landing-page-header.component.html'
+  selector: 'date-location',
+  templateUrl: './date-location.component.html',
+  styleUrl: './date-location.component.css'
 })
-export default class LandingPageHeader implements OnInit {
+export class DateLocation {
   time = new Date();
-  location: string = 'No Location Set';
+  location: string = 'Boston, MA';
   newRunIsLoading = false;
   mobileThreshold = 1070;
   isMobile = window.innerWidth < this.mobileThreshold;
 
-  constructor(private storage: Storage) {}
+  constructor(
+    private storage: Storage,
+    private serverService: APIService,
+    private messageService: MessageService
+  ) {}
 
   ngOnInit() {
+    this.onStartNewRun = () => {
+      const runsQueryResponse = this.serverService.query(() => startNewRun());
+      runsQueryResponse.isLoading.subscribe((isLoading: boolean) => {
+        this.newRunIsLoading = isLoading;
+      });
+      runsQueryResponse.error.subscribe((error: Error) => {
+        this.messageService.add({ severity: 'error', summary: 'Error', detail: error.message });
+      });
+    };
+
     setInterval(() => {
       this.time = new Date();
     }, 1000);
