--- angular-client/src/pages/landing-page/components/landing-page-header/landing-page-header.component.css
@@ -1,5 +1,0 @@
-.header-bar-container {
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-}
