--- angular-client/src/pages/landing-page/components/landing-page-header/landing-page-header.component.html
@@ -1,17 +1,0 @@
-<div style="width: 100%; height: 160px; display: grid; column-gap: 10px">
-  <div style="grid-column-start: 1; grid-column-end: 3; height: 100%">
-    <div style="width: 100%; height: 100%">
-      <typography
-        variant="xx-large-title"
-        [content]="time | date: 'HH:mm:ss'"
-        additionalStyles="font-size: 7rem; text-align: center-bottom"
-      />
-    </div>
-  </div>
-  <div style="grid-column-start: 3; grid-column-end: 4">
-    <landing-page-date-location style="height: 100%; width: 100%" />
-  </div>
-  <landing-page-connection-display style="height: 100%; grid-column-start: 4; grid-column-end: 5" />
-  <landing-page-current-run-display style="height: 100%; grid-column-start: 5; grid-column-end: 6" />
-  <landing-page-viewer-display style="height: 100%; grid-column-start: 6; grid-column-end: 7" />
-</div>
