--- angular-client/src/pages/landing-page/components/landing-page-latency-display/landing-page-latency-display.component.css
@@ -1,23 +1,0 @@
-.latency-display {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-}
-
-.latency-display-body {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0px;
-  height: fit-content;
-}
-
-.latency-display-dot {
-  height: 20px;
-  width: 20px;
-  border-radius: 50%;
-  background-color: inherit;
-  margin-right: 10px;
-}
