+++ angular-client/src/pages/landing-page/components/viewer-display/viewer-display.component.css
