+++ angular-client/src/pages/landing-page/landing-page.component.html
@@ -17,18 +17,18 @@
           />
         </div>
         <div style="display: flex; grid-column-start: 3; grid-column-end: 4">
-          <landing-page-date-location style="height: 100%; width: 100%" />
+          <date-location style="height: 100%; width: 100%" />
         </div>
         <div style="display: flex; grid-column-start: 4; grid-column-end: 5">
-          <landing-page-connection-display style="height: 100%; width: 100%" />
+          <connection-display style="height: 100%; width: 100%" />
         </div>
 
         <div style="display: flex; grid-column-start: 5; grid-column-end: 6">
-          <landing-page-current-run-display style="height: 100%; width: 100%" />
+          <current-run-display style="height: 100%; width: 100%" />
         </div>
 
         <div style="display: flex; grid-column-start: 6; grid-column-end: 7">
-          <landing-page-viewer-display style="width: 100%; height: 100%" />
+          <viewer-display style="width: 100%; height: 100%" />
         </div>
 
         <div style="display: flex; grid-column-start: 1; grid-column-end: 3">
